Fix skipping of deployment (11.)
'Twas not working. The first issue was that I had not negated the
skip_deploy, but even then, trying to refer to the environment through
github.env as suggested did not work. For the variable to be available
on the job-level (rather than step-level) "if", one approach is to
set it in a separate job which the necessary job then depends on and
uses that job's outputs.