Add test workflow for testing discord webhook (11.)
I tried the recommended action, but it complains about not being
able to parse the returned content. Seems to be given HTML while
it tries to parse JSON. Not sure whether it's absolutely necessary
that JSON is received, so trying another action to see whether it's
how it works the response. Then again, no message did pop up in Discord,
which I guess if the POST succeeded such that a response was sent back, as
was the case, you'd think the info went through?