Switch to using different webhook action (11.)
With a little testing, this one actually seemed to succeed. Granted,
I didn't actually try the other one more than once, but that one did
not work on first try and this did, so there we go. This is much more
of a bare bones one, but whatever, it's more about knowing how to
work with webhooks anyway. Could also be more informative with
the messages, but kind of the same point.